Feature/add statistics database insights (#22)
Adds a database to track the player's daily data, i.e. time spent, number of new nouns left and number of nouns revised.

Note that the day is determined by user's device, thus this could be abused to be set in the past/future. Moreover, a day is determined using UTC, thus not user-friendly for those in Asia/Westcoast US, however, for POC, this is acceptable.

Commits:
* Rename PlayerData to PlayerNounData.
* Add lastSeen, isLearned to PlayerNounData.
* Add PlayerDailyData.
* Add IPlayerNounDataDatabase.
* Update IPlayerDataService to use IPlayerNounDataDatabase.
* Add IService.
* Rename _getPlayerData to _getPlayerNounData and variables accordingly.
* Add IPlayerDailyDataDatabase.
* Add DateTimeUtils.
* Add ensureDataExists, updateTimeSpent to IPlayerDailyDataDatabase.
* Update IPlayerDataService to use IPlayerDailyDataDatabase.
* Update GameScreenStore to track time spent playing.